refactor(agent): extract dispatch-batching + repeat-circuit-breaker from dispatcher.ts (#361)
